Add files via upload
In this program the factorial of a number is calculate. It initializes a variable fact to 1, then multiplies it by every integer from 1 to a using a for-loop.
The final factorial value is returned to the main() method, where it is printed on the screen. The input is taken from the user using the Scanner class.